+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 1.6.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix field logic.
+  [cekk]
 
 
 1.6.1 (2022-10-27)