+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 
 - Fix field logic.
   [cekk]
-
+- Set proper value in legislature field.
+  [cekk]
 
 1.6.1 (2022-10-27)
 ------------------